+++ src/commands/handlers/scene/index.ts
@@ -1,3 +1,4 @@
 export { getSceneHandler } from './GetSceneHandler';
 export { getScenesListHandler } from './GetScenesListHandler';
 export { activateSceneHandler } from './ActivateSceneHandler';
+export { captureSceneHandler } from './CaptureSceneHandler';